Split widget AppIntent types for host-target membership
